requirement.txt
django-bootstrap3
pillow
django-pipline
requests
